# perf(gateway): stable round robin