@b0xtch 2023-03-13
I wonder why req/s is very low if all its doing is routing requests.